pyproject.toml
[project]
name = "full-stack-ai-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "ipykernel>=7.1.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]
